--HG--
branch : spec_SFEruby-mini_portile2

+++ specs/SFEperl-cpan-meta-check.spec
@@ -4,7 +4,8 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
 %define include_executable 0
 
 %define cpan_name CPAN-Meta-Check
@@ -33,6 +34,12 @@ BuildRequires:    library/perl-5/cpan-meta-584
 BuildRequires:    library/perl-5/extutils-makemaker-584
 BuildRequires:    library/perl-5/test-deep-584
 BuildRequires:    library/perl-5/test-simple-584
+%if %{enable_test}
+BuildRequires:    library/perl-5/cpan-meta-584
+BuildRequires:    library/perl-5/cpan-meta-requirements-584
+BuildRequires:    library/perl-5/exporter-584
+BuildRequires:    library/perl-5/module-metadata-584
+%endif
 Requires:         runtime/perl-584 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/cpan-meta-584
@@ -53,6 +60,10 @@ BuildRequires:    library/perl-5/cpan-meta-510
 BuildRequires:    library/perl-5/extutils-makemaker-510
 BuildRequires:    library/perl-5/test-deep-510
 BuildRequires:    library/perl-5/test-simple-510
+BuildRequires:    library/perl-5/cpan-meta-510
+BuildRequires:    library/perl-5/cpan-meta-requirements-510
+BuildRequires:    library/perl-5/exporter-510
+BuildRequires:    library/perl-5/module-metadata-510
 Requires:         runtime/perl-510 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/cpan-meta-510
@@ -73,6 +84,12 @@ BuildRequires:    library/perl-5/cpan-meta-512
 BuildRequires:    library/perl-5/extutils-makemaker-512
 BuildRequires:    library/perl-5/test-deep-512
 BuildRequires:    library/perl-5/test-simple-512
+%if %{enable_test}
+BuildRequires:    library/perl-5/cpan-meta-512
+BuildRequires:    library/perl-5/cpan-meta-requirements-512
+BuildRequires:    library/perl-5/exporter-512
+BuildRequires:    library/perl-5/module-metadata-512
+%endif
 Requires:         runtime/perl-512 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/cpan-meta-512
@@ -93,6 +110,13 @@ BuildRequires:    library/perl-5/cpan-meta-516
 BuildRequires:    library/perl-5/extutils-makemaker-516
 BuildRequires:    library/perl-5/test-deep-516
 BuildRequires:    library/perl-5/test-simple-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+BuildRequires:    library/perl-5/cpan-meta-516
+BuildRequires:    library/perl-5/cpan-meta-requirements-516
+BuildRequires:    library/perl-5/exporter-516
+BuildRequires:    library/perl-5/module-metadata-516
+%endif
 Requires:         runtime/perl-516 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/cpan-meta-516
@@ -104,30 +128,36 @@ Requires:         library/perl-5/module-metadata-516
 Verify requirements in a CPAN::Meta object
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Verify requirements in a CPAN::Meta object
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/cpan-meta-520
-BuildRequires:    library/perl-5/extutils-makemaker-520
-BuildRequires:    library/perl-5/test-deep-520
-BuildRequires:    library/perl-5/test-simple-520
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/cpan-meta-522
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/test-deep-522
+BuildRequires:    library/perl-5/test-simple-522
+%if %{enable_test}
+BuildRequires:    library/perl-5/cpan-meta-522
+BuildRequires:    library/perl-5/cpan-meta-requirements-522
+BuildRequires:    library/perl-5/exporter-522
+BuildRequires:    library/perl-5/module-metadata-522
+%endif
+Requires:         runtime/perl-522 = *
 Requires:         library/perl-5/%{ips_cpan_name}
-Requires:         library/perl-5/cpan-meta-520
-Requires:         library/perl-5/cpan-meta-requirements-520
-Requires:         library/perl-5/exporter-520
-Requires:         library/perl-5/module-metadata-520
+Requires:         library/perl-5/cpan-meta-522
+Requires:         library/perl-5/cpan-meta-requirements-522
+Requires:         library/perl-5/exporter-522
+Requires:         library/perl-5/module-metadata-522
 
-%description 520
+%description 522
 Verify requirements in a CPAN::Meta object
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -140,8 +170,12 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ x${test} = 'xwithout_test' ] || make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
@@ -156,7 +190,7 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ x${test} = 'xwithout_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build clean
 }
@@ -173,7 +207,7 @@ modify_bin_dir() {
     then
         for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
         do
-            sed -i.bak -e "s/\/usr\/bin\/env ruby/\/usr\/perl5\/${perl-ver}\/bin\/ruby/" ${i}
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
             [ -f ${i}.bak] || rm -f ${i}.bak
         done
     fi
@@ -230,8 +264,8 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
@@ -292,17 +326,18 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Thu Jun 01 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- build package for perl-522 instead of perl-520
 * Sat Dec 05 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - initial commit
