+++ specs/SFEperl-file-path.spec
@@ -1,17 +1,20 @@
 %include Solaris.inc
 
 %define build584 0
+%define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
 %define include_executable 0
 
 %define cpan_name File-Path
 %define sfe_cpan_name file-path
+%define ips_cpan_name file-path
 
 Summary:               Create or remove directory trees
 Name:                  SFEperl-%{sfe_cpan_name}
-IPS_package_name:      library/perl-5/%{sfe_cpan_name}
+IPS_package_name:      library/perl-5/%{ips_cpan_name}
 Version:               2.12
 IPS_component_version: 2.12
 License:               perl_5
@@ -24,29 +27,54 @@ Create or remove directory trees
 
 %if %{build584}
 %package 584
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-584
+IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Create or remove directory trees
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-584
+BuildRequires:    library/perl-5/pathtools-584
+%endif
 Requires:         runtime/perl-584 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-584
 Requires:         library/perl-5/pathtools-584
-Requires:         library/perl-5/selectsaver-584
 
 %description 584
 Create or remove directory trees
 %endif
 
+%if %{build510}
+%package 510
+IPS_package_name: library/perl-5/%{ips_cpan_name}-510
+Summary:          Create or remove directory trees
+BuildRequires:    runtime/perl-510 = *
+BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/carp-510
+BuildRequires:    library/perl-5/pathtools-510
+Requires:         runtime/perl-510 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/carp-510
+Requires:         library/perl-5/pathtools-510
+
+%description 510
+Create or remove directory trees
+%endif
+
 %if %{build512}
 %package 512
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-512
+IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Create or remove directory trees
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-512
+BuildRequires:    library/perl-5/pathtools-512
+%endif
 Requires:         runtime/perl-512 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-512
 Requires:         library/perl-5/pathtools-512
-Requires:         library/perl-5/selectsaver-512
 
 %description 512
 Create or remove directory trees
@@ -54,38 +82,47 @@ Create or remove directory trees
 
 %if %{build516}
 %package 516
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-516
+IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Create or remove directory trees
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-516
+BuildRequires:    library/perl-5/pathtools-516
+%endif
 Requires:         runtime/perl-516 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-516
 Requires:         library/perl-5/pathtools-516
-Requires:         library/perl-5/selectsaver-516
 
 %description 516
 Create or remove directory trees
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Create or remove directory trees
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-Requires:         runtime/perl-520 = *
-Requires:         library/perl-5/carp-520
-Requires:         library/perl-5/pathtools-520
-Requires:         library/perl-5/selectsaver-520
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-522
+BuildRequires:    library/perl-5/pathtools-522
+%endif
+Requires:         runtime/perl-522 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/carp-522
+Requires:         library/perl-5/pathtools-522
 
-%description 520
+%description 522
 Create or remove directory trees
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -98,8 +135,12 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ ${test} == 'without_test' ] || make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
@@ -114,35 +155,70 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ ${test} == 'without_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
+    ${bindir}/perl ./Build clean
 }
 
 modify_bin_dir() {
-  perl_ver=$1
-  if [ -d $RPM_BUILD_ROOT/usr/bin ]
-  then
-    [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
-    mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
-  fi
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/bin ]
+    then
+      [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+      mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
+    fi
+      
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
+    then
+        for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
+        do
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
+            [ -f ${i}.bak] || rm -f ${i}.bak
+        done
+    fi
+}
+
+modify_man_dir() {
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man ]
+    then
+        if [ -d $RPM_BUILD_ROOT%{_datadir}/man ]
+        then
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        else
+            mkdir -p $RPM_BUILD_ROOT%{_datadir}
+            mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        fi
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
+    fi
 }
 
 build_for() {
-  if [ -f Makefile.PL ];
-  then
-    build_with_makefile.pl_for $*
-  elif [ -f Build.PL ];
+  if [ -f Build.PL ];
   then
     build_with_build.pl_for $*
+  elif [ -f Makefile.PL ];
+  then
+    build_with_makefile.pl_for $*
   fi
 
-    modify_bin_dir $*
+  modify_bin_dir $*
+  modify_man_dir $*
 }
 
 # To build without test, pass 'without_test' to build_for commaond.
 # like 'build_for version without_test'
 %if %{build584}
 build_for 5.8.4
+%endif
+
+%if %{build510}
+build_for 5.10
 %endif
 
 %if %{build512}
@@ -153,14 +229,14 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
-mkdir -p $RPM_BUILD_ROOT%{_datadir}
 if [ -d $RPM_BUILD_ROOT%{_prefix}/man ]
 then
+    mkdir -p $RPM_BUILD_ROOT%{_datadir}
     mv $RPM_BUILD_ROOT%{_prefix}/man $RPM_BUILD_ROOT%{_datadir}
 fi
 if [ -d $RPM_BUILD_ROOT%{_datadir}/man/man3 ]
@@ -185,6 +261,16 @@ rm -rf %{buildroot}
 %endif
 %endif
 
+%if %{build510}
+%files 510
+%defattr(0755,root,bin,-)
+%dir %attr (0755, root, sys) /usr
+/usr/perl5/vendor_perl/5.10
+%if %{include_executable}
+/usr/perl5/5.1.0
+%endif
+%endif
+
 %if %{build512}
 %files 512
 %defattr(0755,root,bin,-)
@@ -205,17 +291,20 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Wed May 24 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- fix requires. selectsaver is included in perl
+* Wed Apr 26 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- package for perl-522 is added and for perl-520 is obsolete
 * Wed Nov 11 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - initial commit
